refactor : CommonController, UserRepository의 패키지 이동
