+++ src/main/kotlin/com/flab/ticketing/auth/service/AuthService.kt
@@ -2,7 +2,6 @@ package com.flab.ticketing.auth.service
 
 import com.flab.ticketing.auth.dto.UserRegisterDto
 import com.flab.ticketing.auth.exception.AuthErrorInfos
-import com.flab.ticketing.auth.repository.UserRepository
 import com.flab.ticketing.auth.utils.EmailCodeGenerator
 import com.flab.ticketing.auth.utils.EmailSender
 import com.flab.ticketing.auth.utils.EmailVerifier
@@ -11,6 +10,7 @@ import com.flab.ticketing.common.exception.InvalidValueException
 import com.flab.ticketing.common.exception.NotFoundException
 import com.flab.ticketing.common.utils.NanoIdGenerator
 import com.flab.ticketing.user.entity.User
+import com.flab.ticketing.user.entity.repository.UserRepository
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
