+++ src/test/kotlin/com/flab/ticketing/auth/service/AuthServiceTest.kt
@@ -1,7 +1,7 @@
 package com.flab.ticketing.auth.service
 
 import com.flab.ticketing.auth.dto.UserRegisterDto
-import com.flab.ticketing.auth.repository.UserRepository
+import com.flab.ticketing.user.entity.repository.UserRepository
 import com.flab.ticketing.auth.utils.EmailCodeGenerator
 import com.flab.ticketing.auth.utils.EmailSender
 import com.flab.ticketing.auth.utils.EmailVerifier
